# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Heuvera",
    "version": "1.0.0",
    "description": "Heuvera API documentation with Swagger"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/v1/bookings": {
      "get": {
        "summary": "Get all bookings with analytics",
        "description": "Admins can retrieve all bookings with pagination, total bookings, and revenue stats.",
        "tags": ["Bookings"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of bookings per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings with analytics"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/v1/hosts": {
      "get": {
        "summary": "Get all users who are hosts",
        "description": "Fetches all hosts with their details, including verification status and listings count.",
        "tags": ["Hosts"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of hosts per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of hosts with their details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/v1/hosts/status": {
      "patch": {
        "summary": "Approve a host",
        "description": "Admins can approve users as hosts by setting isHostApproved to true.",
        "tags": ["Hosts"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Host approved successfully"
          },
          "400": {
            "description": "User ID is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/v1/listings": {
      "get": {
        "summary": "Get all listings with pagination",
        "description": "Admins can retrieve a paginated list of all property listings along with the total count.",
        "tags": ["Listings"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of listings per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of listings with total count"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/v1/reviews": {
      "get": {
        "summary": "Get all reviews and ratings",
        "description": "Admins can fetch all reviews along with average rating analytics.",
        "tags": ["Reviews"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of reviews per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews with analytics"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}