# src/app/api/admin/v1/reviews/route.ts
import { NextRequest, NextResponse } from "next/server";
import { prisma } from "@heuvera/lib/prisma";
import { authMiddleware } from "@heuvera/lib/admin/middleware";

/**
 * @swagger
 * /api/admin/v1/reviews:
 *   get:
 *     summary: Get all reviews and ratings
 *     description: Admins can fetch all reviews along with average rating analytics.
 *     tags: [Reviews]
 *     security:
 *       - BearerAuth: []
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *           example: 1
 *         description: The page number (default is 1)
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *           example: 10
 *         description: Number of reviews per page (default is 10)
 *     responses:
 *       200:
 *         description: List of reviews with analytics
 *       401:
 *         description: Unauthorized
 *       500:
 *         description: Internal Server Error
 */
export async function GET(req: NextRequest): Promise<Response> {
  try {
    const adminResult = authMiddleware(req);

    if (typeof adminResult !== "object" || adminResult instanceof String) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    if (adminResult instanceof NextResponse) {
      return adminResult;
    }

    const { searchParams } = new URL(req.url);
    const page = parseInt(searchParams.get("page") ?? "1", 10);
    const limit = parseInt(searchParams.get("limit") ?? "10", 10);
    const skip = (page - 1) * limit;

    const [reviews, totalReviews, averageRating] = await prisma.$transaction([
      prisma.review.findMany({
        skip,
        take: limit,
        orderBy: { createdAt: "desc" },
        select: {
          id: true,
          userId: true,
          listingId: true,
          rating: true,
          comment: true,
          createdAt: true,
        },
      }),
      prisma.review.count(),
      prisma.review.aggregate({ _avg: { rating: true } }),
    ]);

    return NextResponse.json({
      totalReviews,
      averageRating: averageRating._avg.rating ?? 0,
      reviews,
    });
  } catch (error) {
    console.error("Error fetching reviews:", error);
    return NextResponse.json(
      { error: "Internal Server Error" },
      { status: 500 },
    );
  }
}